src/requirements.txt
streamlit 
transformers
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
python-docx
spacy
docx2txt
cupy-cuda111
sentence-transformers
en_core_web_sm-3.7.1-py3-none-any.whl
pymupdf
nltk
pyspellchecker

